+++ packages/pn-pa-webapp/src/components/NewNotification/PreliminaryInformations.tsx
@@ -78,7 +78,7 @@ const PreliminaryInformations = ({ notification, onConfirm }: Props) => {
 
   const initialValues = useCallback(() => {
     const additionalLang = additionalLanguages?.length > 0 ? additionalLanguages[0] : undefined;
-
+     
     return {
       paProtocolNumber: notification.paProtocolNumber || '',
       subject: notification.subject || '',
@@ -97,7 +97,12 @@ const PreliminaryInformations = ({ notification, onConfirm }: Props) => {
 
   const validationSchema = yup.object({
     paProtocolNumber: requiredStringFieldValidation(tc, 256),
-    subject: requiredStringFieldValidation(tc, 134, 10),
+    subject: yup.string()
+     .when('lang',{
+      is: NewNotificationLangOther,
+      then: requiredStringFieldValidation(tc, 66, 10),
+      otherwise: requiredStringFieldValidation(tc, 134, 10)
+    }),
     senderDenomination: yup
       .string()
       .required(`${t('sender-denomination')} ${tc('required')}`)
@@ -122,7 +127,7 @@ const PreliminaryInformations = ({ notification, onConfirm }: Props) => {
     }),
     additionalSubject: yup.string().when('lang', {
       is: NewNotificationLangOther,
-      then: requiredStringFieldValidation(tc, 134, 10),
+      then: requiredStringFieldValidation(tc, 66, 10),
     }),
     additionalAbstract: yup.string().when('lang', {
       is: NewNotificationLangOther,
@@ -153,6 +158,11 @@ const PreliminaryInformations = ({ notification, onConfirm }: Props) => {
   };
 
   const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const value = e.target.value;
+    if(value === 'it'){
+      void formik.setValues({...formik.values, additionalLang:'', additionalSubject: ''},false);
+      void formik.setFieldTouched('additionalSubject',false);
+    }
     formik.handleChange(e);
   };