+++ packages/pn-pa-webapp/src/utility/AppError/PAAppErrorFactory.ts
@@ -8,6 +8,7 @@ import {
 import { GenericInvalidParameterAppError } from './GenericInvalidParameterAppError';
 import { GenericInvalidParameterDuplicateAppError } from './GenericInvalidParameterDuplicateAppError';
 import { ServerResponseErrorCode } from './types';
+import { GenericInvalidParameterTaxonomyCodeAppError } from './GenericInvalidParameterTaxonomyCodeAppError';
 
 export class PAAppErrorFactory extends AppErrorFactory {
   private translateFunction: (path: string, ns: string) => string;
@@ -25,6 +26,8 @@ export class PAAppErrorFactory extends AppErrorFactory {
         return new GenericInvalidParameterAppError(error, this.translateFunction);
       case ServerResponseErrorCode.PN_GENERIC_INVALIDPARAMETER_DUPLICATED:
         return new GenericInvalidParameterDuplicateAppError(error, this.translateFunction);
+      case ServerResponseErrorCode.PN_GENERIC_INVALIDPARAMETER_TAXONOMYCODE:
+        return new GenericInvalidParameterTaxonomyCodeAppError(error, this.translateFunction);
       default:
         return new UnknownAppError(error);
     }