+++ packages/pn-pa-webapp/src/utility/notification.utility.ts
@@ -228,7 +228,7 @@ export function getDuplicateValuesByKeys<T>(
 
 const concatAdditionalContent = (content?: string, additionalContent?: string): string => {
   if (content && additionalContent) {
-    return `${content} • ${additionalContent}`;
+    return `${content}|${additionalContent}`;
   }
   return content || additionalContent || '';
 };