+++ packages/pn-personafisica-webapp/src/redux/contact/__test__/reducers.test.ts
@@ -30,6 +30,7 @@ import {
   resetState,
   setExternalEvent,
 } from '../reducers';
+import { sortAddresses } from '../../../utility/contacts.utility';
 
 const initialState = {
   loading: false,
@@ -63,7 +64,7 @@ describe('Contacts redux state tests', () => {
     mock.onGet('/bff/v1/addresses').reply(200, digitalAddresses);
     const action = await store.dispatch(getDigitalAddresses());
     expect(action.type).toBe('getDigitalAddresses/fulfilled');
-    expect(action.payload).toEqual(digitalAddresses);
+    expect(action.payload).toEqual(sortAddresses(digitalAddresses));
   });
 
   it('Should be able to update the digital address with legal value (pec to verify)', async () => {