+++ packages/pn-personagiuridica-webapp/src/utility/__test__/contacts.utility.test.ts
@@ -6,6 +6,7 @@ import { SelectedAddresses } from '../../redux/contact/reducers';
 import {
   contactAlreadyExists,
   removeAddress,
+  sortAddresses,
   specialContactsAvailableAddressTypes,
   updateAddressesList,
 } from '../contacts.utility';
@@ -161,6 +162,106 @@ describe('Contacts utility test', () => {
     ]);
   });
 
+  it('test sortAddresses function - legal contacts', () => {
+    const defaultPecAddress = digitalAddresses.find(
+      addr => addr.channelType === ChannelType.PEC && addr.senderId === 'default'
+    )!;
+    const defaultSercqAddress = digitalAddressesSercq.find(
+      addr => addr.channelType === ChannelType.SERCQ_SEND
+    )!;
+
+    // test with two default addresses
+    let addresses: Array<DigitalAddress> = [
+      {
+        ...defaultPecAddress,
+        value: '',
+        pecValid: false
+      },
+      defaultSercqAddress
+    ];
+    let sortedAddresses = [addresses[1], addresses[0]];
+
+    expect(sortedAddresses).toStrictEqual(sortAddresses(addresses));
+
+    addresses = [
+      defaultSercqAddress,
+      {
+        ...defaultPecAddress,
+        value: '',
+        pecValid: false
+      }
+    ];
+    sortedAddresses = [addresses[0], addresses[1]];
+
+    expect(sortedAddresses).toStrictEqual(sortAddresses(addresses));
+
+    // test with one default and one non default addresses
+    addresses = [
+      {
+        ...defaultPecAddress,
+        senderId: 'comune-milano'
+      },
+      defaultSercqAddress
+    ];
+    sortedAddresses = [addresses[1], addresses[0]];
+
+    expect(sortedAddresses).toStrictEqual(sortAddresses(addresses));
+
+    // test with two non default addresses
+    addresses = [
+      {
+        ...defaultPecAddress,
+        senderId: 'comune-milano'
+      },
+      {
+        ...defaultSercqAddress,
+        senderId: 'comune-milano'
+      }
+    ];
+    sortedAddresses = [addresses[1], addresses[0]];
+
+    expect(sortedAddresses).toStrictEqual(sortAddresses(addresses));
+  });
+
+  it('test sortAddresses function - courtesy contacts', () => {
+    const mailAddress = digitalAddresses.find(addr => addr.channelType === ChannelType.EMAIL)!;
+    const smsAddress = digitalAddresses.find(addr => addr.channelType === ChannelType.SMS)!;
+
+    let addresses = [mailAddress, smsAddress];
+    let sortedAddresses = [addresses[0], addresses[1]];
+
+    expect(sortedAddresses).toStrictEqual(sortAddresses(addresses));
+
+    // check the other way around
+    addresses = [smsAddress, mailAddress];
+    sortedAddresses = [addresses[1], addresses[0]];
+
+    expect(sortedAddresses).toStrictEqual(sortAddresses(addresses));
+  });
+
+  it('test sortAddresses function - both legal and courtesy', () => {
+    const defaultPecAddress = digitalAddresses.find(
+      addr => addr.channelType === ChannelType.PEC && addr.senderId === 'default'
+    )!;
+    const defaultSercqAddress = digitalAddressesSercq.find(
+      addr => addr.channelType === ChannelType.SERCQ_SEND
+    )!;
+    const mailAddress = digitalAddresses.find(addr => addr.channelType === ChannelType.EMAIL)!;
+    const smsAddress = digitalAddresses.find(addr => addr.channelType === ChannelType.SMS)!;
+    
+    // test sortAddress with PEC and SMS
+    let addresses = [smsAddress, defaultPecAddress];
+    let sortedAddresses = [addresses[1], addresses[0]];
+
+    expect(sortedAddresses).toStrictEqual(sortAddresses(addresses));
+    
+    // test sortAddress with SERCQ and EMAIL
+    addresses = [mailAddress, defaultSercqAddress];
+    sortedAddresses = [addresses[1], addresses[0]];
+
+    expect(sortedAddresses).toStrictEqual(sortAddresses(addresses));
+  });
+
   it('test updateAddressesList function, new address', () => {
     const newAddress: DigitalAddress = {
       addressType: AddressType.LEGAL,
@@ -177,7 +278,7 @@ describe('Contacts utility test', () => {
       newAddress
     );
 
-    expect(digitalAddresses).toStrictEqual([...previousDigitalAddresses, newAddress]);
+    expect(digitalAddresses).toStrictEqual(sortAddresses([...previousDigitalAddresses, newAddress]));
   });
 
   it('test updateAddressesList function, existing courtesy address', () => {